services/inventory/requirements.txt
Django>=4.2,<5
django-ninja>=1.0
django-cors-headers>=4.3
gunicorn>=21.0
